A quick pass at integrating vertx-lib
The basic idea is to put all the stuff that is currently in MainVerticle
in a new AuthorizeApi class that inherits from RouterCreator. Then
use the folio-vertx-lib way of mouting the routerCreators.

This compiles but needs more work before testing.

Some things that need to be done to make it actually work:
1. Add the body handler route to AuthorizeApi.
2. Actually see it work and unit tests pass.
3. Figure out to do with some of the tests that I commented out.
